rename album bug corrected
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@7080 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ browserController.m
@@ -9924,8 +9924,9 @@ static BOOL needToRezoom;
 			[NSApp endSheet: newAlbum];
 			[newAlbum orderOut: self];
 			
-			if( result == NSRunStoppedResponse)	{
-				long				i = 2;
+			if( result == NSRunStoppedResponse)
+			{
+				int i = 2;
 				
 				if( [[newAlbumName stringValue] isEqualToString: [album valueForKey:@"name"]] == NO)
 				{
